+++ src/const/Enum.js
@@ -70,7 +70,7 @@ const Enum = {
   STORY_2A_CLAIM_BLUE: 4,
   STORY_2B_PLACE_FLAG: 5,
   STORY_2C_LEAVE_FOREST: 6,
-  STORY_3_DESTROY_ROSE: 7,
+  STORY_3_CHASE_REDFACE: 7,
   STORY_4_CLAIM_STORM: 8,
 
   //  Dialogue
