+++ src/const/Icon.js
@@ -1,5 +1,7 @@
 const Icon = {
   SPEECH: 26,
-  ALARM: 47
+  ALARM: 47,
+  ANGER: 219,
+  SPEAR: 473,
 }
 export default Icon;
